adds workaround solving space enconding during test execution
When CI tests are wrapped in scripts for execution in certain environments,
spaces obtained as JVM resources (useful for safe relative-path resolution)
are encoded as URLs.  While useful and largely portable, when executions
are wrapped in scripts, shell-executing environments fail to read these
when searching for local paths.  This causes test resources to not be found.

See http://jlorenzen.blogspot.com/2007/06/proper-way-to-access-file-resources-in.html
(I8UMuch comments, on September 11th, 2009)

This hack obtains the URL in universal format, and manually re-writes the
String output to re-replace the escaped spaces.  IMHO, this will probably
cause problems for other encoded.  A better solution would probably be
to decode the URL entirely to String format to feed to the file.

I've added a TODO to create a re-usable getResourceAsFile() method.