makes schedule resource for integration test shared
The schedule resource constructor now (appropriately) instantiates (real)
schedules from UPL and performs OCR on object construction.  This was making
tests extremely slow, and taxing resources.

It may be better to obtain the schedule from a Factory class, which would
return all similar schedules as a single in-memory object.  As a stopgap,
it may be advantageous to make the Resources singleton objects (though I
dislike this for reasons I can't as yet put my finger on).

For now, this makes the integration tests run much faster.

Since methods instatiating the objects also make calls to external resources,
the next commits will move all such tests to the integration test tree.