disables broken integration tests
The Ultimate Pool League Web site has changed the way in which
they publish results, which currently causes OCR failures during
what we in this project refer to as "integration" tests.

In an effort to re-enable Pull reqeust building via Travis CI,
these tests are disabled.  The Trello board will need to be modified
to reflect our new approach to building the project.